requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

# MQTT
paho-mqtt==1.6.1

# Modbus (optional — for full Modbus server)
pymodbus==3.6.8

# OPC-UA (optional — for full OPC-UA server)
opcua==0.98.13

# InfluxDB
influxdb-client==1.43.0

# TimescaleDB / PostgreSQL
psycopg2-binary==2.9.9

# Testing
pytest==8.2.0
pytest-cov==5.0.0
pytest-asyncio==0.23.6
httpx==0.27.0
